Changed loop vars from int to size_t.

+++ src/GameOfLife.cpp
@@ -81,8 +81,8 @@ int GameOfLife::countNeighbors(const int x, const int y) {
 	ydomain[1] = y;
 	ydomain[2] = (y == height - 1 ? 0 : y + 1);
 	
-	for(int i = 0; i < 3; ++i) {
-		for(int j = 0; j < 3; ++j) {
+	for(size_t i = 0; i < 3; ++i) {
+		for(size_t j = 0; j < 3; ++j) {
             if (!(xdomain[i] == x && ydomain[j] == y)) {
                 if (grid[xdomain[i]][ydomain[j]]) {
                     ++neighbors;
